+++ CHANGELOG.md
@@ -35,6 +35,8 @@ All user visible changes to this project will be documented in this file. This p
     - Chats tab:
         - Redesigned chats appearing animation. ([#1514])
         - Redesigned unread messages counter. ([#1512], [#1507])
+    - Menu tab:
+        - Redesigned background color. [[#1522], [#1524]]
 
 ### Fixed
 
@@ -49,6 +51,8 @@ All user visible changes to this project will be documented in this file. This p
 [#1507]: /../../issues/1507
 [#1512]: /../../pull/1512
 [#1514]: /../../pull/1514
+[#1522]: /../../pull/1522
+[#1524]: /../../issues/1524